refactor: prepare for mocking out cost and effect of gameActions
