+++ src/state/gameActions/findGameActionById.ts
@@ -1,0 +1,6 @@
+import { GameAction } from './types';
+import { GameActionId, gameActions } from './gameActions';
+
+export function findGameActionById(gameActionId: GameActionId): GameAction {
+  return gameActions.find((gameAction) => gameActionId === gameAction.id)!;
+}
