+++ src/state/gameActions/getAvailableGameActions.ts
@@ -1,0 +1,17 @@
+import { GameAction } from './types';
+import { GameActionId, gameActions } from './gameActions';
+
+export function getAvailableGameActions(
+  currentRound: number,
+  finishedActionIds: GameActionId[],
+  selectedGameActionIds: GameActionId[],
+): GameAction[] {
+  return gameActions.filter((gameAction) =>
+    gameAction.available(
+      currentRound,
+      finishedActionIds,
+      selectedGameActionIds,
+      gameAction.id,
+    ),
+  );
+}
