+++ src/state/gameActions/getCost.ts
@@ -1,0 +1,5 @@
+import { GameAction } from './types';
+
+export function getCost(gameAction: GameAction): number {
+  return gameAction.cost;
+}
