+++ src/state/gameActions/index.ts
@@ -1,6 +1,9 @@
 import { GameActionId as GameActionIdT } from './gameActions';
 import { GameAction as GameActionT } from './types';
-export * from './derivers';
+export { getAvailableGameActions } from './getAvailableGameActions';
+export { findGameActionById } from './findGameActionById';
+export { getEffect } from './getEffect';
+export { getCost } from './getCost';
 
 export type GameActionId = GameActionIdT;
 export type GameAction = GameActionT;
